+++ src/main/java/com/mindera/school/spaceshiprent/service/auth/AuthService.java
@@ -1,0 +1,9 @@
+package com.mindera.school.spaceshiprent.service.auth;
+
+import com.mindera.school.spaceshiprent.dto.user.CredentialsDto;
+import com.mindera.school.spaceshiprent.dto.user.LoginDto;
+
+public interface AuthService {
+
+    LoginDto login(CredentialsDto credentials);
+}
